updated restful methods in service and respository for habit to limit only owner can do crud

+++ src/main/java/com/capstone/habitbuilder/habit/HabitService.java
@@ -3,21 +3,41 @@ package com.capstone.habitbuilder.habit;
 import com.capstone.habitbuilder.enjoyer.Enjoyer;
 import com.capstone.habitbuilder.enjoyer.EnjoyerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 public class HabitService {
     private final HabitRepository habitRepository;
     private final EnjoyerRepository enjoyerRepository;
+
     @Autowired
     public HabitService(HabitRepository habitRepository, EnjoyerRepository enjoyerRepository) {
         this.habitRepository = habitRepository;
         this.enjoyerRepository = enjoyerRepository;
     }
-    
+
+    public String authenticateUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication instanceof AnonymousAuthenticationToken)) {
+            String currentUserName = authentication.getName();
+            return currentUserName;
+        } else {
+            return "Can't now find this user";
+        }
+    }
+
     // index - need to change based on userId
     public Iterable<Habit> getHabits(Long enjoyerId) {
+        Enjoyer enjoyer = enjoyerRepository.findById(enjoyerId)
+                .orElseThrow(() -> new IllegalStateException(
+                        "enjoyer with id " + enjoyerId + " does not exists"
+                ));
         return habitRepository.findByEnjoyerId(enjoyerId);
     }
 
@@ -27,7 +47,16 @@ public class HabitService {
                 .orElseThrow(() -> new IllegalStateException(
                         "habit with id " + habitId + " does not exists"
                 ));
-        return habit;
+        Enjoyer enjoyer = enjoyerRepository.findEnjoyerByEmail(authenticateUser())
+                .orElseThrow(() -> new IllegalStateException(
+                        "User with email " + authenticateUser() + " does not exists"
+                ));
+        if (habitRepository.findByIdAndEnjoyer(habitId, enjoyer) != null) {
+            return habit;
+        } else{
+            throw new IllegalStateException(
+                    "habitId " + habitId + " and enjoyerId" + enjoyer.getId() + " does not exists");
+        }
     }
 
     // Create
@@ -42,33 +71,37 @@ public class HabitService {
 
     // Update
     @Transactional
-    public void updateHabit(Long habitId,
-                            String title,
-                            String goal,
-                            String description,
-                            String streak,
-                            Boolean reminder,
-                            Boolean habitBuilt) {
-        Habit habit = habitRepository.findById(habitId)
+    public void updateHabit(Long habitId, Habit habit) {
+        Habit checkHabit = habitRepository.findById(habitId)
                 .orElseThrow(() -> new IllegalStateException(
                         "habit with id " + habitId + " does not exists"
                 ));
-        if (title != null && title.length() > 0 ) {
-            habit.setTitle(title);
-        }
-        if (goal != null && goal.length() > 0 ) {
-            habit.setGoal(goal);
-        }
-        if (description != null && description.length() > 0 ) {
-            habit.setDescription(description);
-        }
-        if (streak != null && streak.length() > 0 ) {
-            habit.setStreak(streak);
-        }
+        Enjoyer enjoyer = enjoyerRepository.findEnjoyerByEmail(authenticateUser())
+                .orElseThrow(() -> new IllegalStateException(
+                        "User with email " + authenticateUser() + " does not exists"
+                ));
 
-        habit.setReminder(reminder);
-        habit.setHabitBuilt(habitBuilt);
-        habitRepository.save(habit);
+        if (habitRepository.findByIdAndEnjoyer(habitId, enjoyer) != null) {
+            if (habit.getTitle() != null && habit.getTitle().length() > 0 ) {
+                habit.setTitle(habit.getTitle());
+            }
+            if (habit.getGoal() != null && habit.getGoal().length() > 0 ) {
+                habit.setGoal(habit.getGoal());
+            }
+            if (habit.getDescription() != null && habit.getDescription().length() > 0 ) {
+                habit.setDescription(habit.getDescription());
+            }
+            if (habit.getStreak() != null && habit.getStreak().length() > 0 ) {
+                habit.setStreak(habit.getStreak());
+            }
+
+            habit.setReminder(habit.getReminder());
+            habit.setHabitBuilt(habit.getHabitBuilt());
+            habitRepository.save(habit);
+        } else{
+            throw new IllegalStateException(
+                    "habitId " + habitId + " and enjoyerId" + enjoyer.getId() + " does not exists");
+        }
     }
 
     // Delete
@@ -78,6 +111,15 @@ public class HabitService {
             throw new IllegalStateException(
                     "habit with id " + habitId + " does not exists");
         }
-        habitRepository.deleteById(habitId);
+        Enjoyer enjoyer = enjoyerRepository.findEnjoyerByEmail(authenticateUser())
+                .orElseThrow(() -> new IllegalStateException(
+                        "User with email " + authenticateUser() + " does not exists"
+                ));
+        if (habitRepository.findByIdAndEnjoyer(habitId, enjoyer) != null) {
+            habitRepository.deleteById(habitId);
+        } else{
+            throw new IllegalStateException(
+                    "habitId " + habitId + " and enjoyerId" + enjoyer.getId() + " does not exists");
+        }
     }
 }
